modifying product delete to check for owner and orders

+++ Bangazon/Controllers/ProductsController.cs
@@ -35,9 +35,9 @@ namespace Bangazon.Controllers
                 .Include(p => p.User).ToListAsync();
             if (!String.IsNullOrEmpty(searchString))
             {
-                 productList = productList.Where(p => p.Title.Contains(searchString)
-                                       || p.Description.Contains(searchString)).ToList();
-            } 
+                productList = productList.Where(p => p.Title.Contains(searchString)
+                                      || p.Description.Contains(searchString)).ToList();
+            }
 
             var applicationDbContext = productList
                 .OrderByDescending(p => p.DateCreated)
@@ -53,6 +53,19 @@ namespace Bangazon.Controllers
                 .Include(p => p.ProductType)
                 .Include(p => p.User)
                 .Where(p => p.ProductTypeId == id);
+            if (products == null)
+            {
+                return NotFound();
+            }
+            return View(await products.ToListAsync());
+        }
+
+        public async Task<IActionResult> MyProducts()
+        {
+            var loggedInUser = await GetCurrentUserAsync();
+            var products = _context.Product
+                .Include(p => p.ProductType)
+                .Where(p => p.UserId == loggedInUser.Id);
             if (products == null)
             {
                 return NotFound();
@@ -141,6 +154,8 @@ namespace Bangazon.Controllers
                 return NotFound();
             }
 
+            ModelState.Remove("UserId");
+
             if (ModelState.IsValid)
             {
                 try
@@ -191,10 +206,33 @@ namespace Bangazon.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var loggedInUser = await GetCurrentUserAsync();
             var product = await _context.Product.FindAsync(id);
-            _context.Product.Remove(product);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
+            if (product.UserId == loggedInUser.Id)
+            {
+                foreach (var item in _context.OrderProduct)
+                {
+                    if (product.ProductId == item.ProductId)
+                    {
+                        //error
+                        return View();
+                    }
+                    else
+                    {
+                        _context.Product.Remove(product);
+                        await _context.SaveChangesAsync();
+                        return RedirectToAction(nameof(Index));
+                    }
+                }
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                //error
+                return View();
+            }
+
         }
 
         private bool ProductExists(int id)
